made Rectangle into a template class to support integer and float implementations

+++ src/display/Texture.cpp
@@ -78,10 +78,7 @@ void Texture::setColor(Uint8 red, Uint8 green, Uint8 blue)
 void Texture::render(const int x, const int y, const Rectangle<int>& clip)
 {
     // Set rendering space and render to screen
-    Rectangle<int> renderQuad = {x,
-                                 y,
-                                 static_cast<float>(_width),
-                                 static_cast<float>(_height)};
+    Rectangle<int> renderQuad = {x, y, _width, _height};
 
     // Set clip rendering dimensions
     if (!clip.isNull()) {
